Changes primary code package directory name to 'pulseblaster'
Previously, source code were in src directory. This does not
result in an installation of code as expected. Either a subfolder
called pulseblaster or changin the top-level folder name to
pulseblaster results in a correct installation.

Adds to .gitignore

Changes the spinapi object inside of spinapi.py to _spinapi in
order to clearly distinguish the namespace from the object.

With the correct install package name, import of modules follows
standard convention. The example code pb_infinte_on_example
is updated accordingly.